added shavian lettering
I'm not sure if this is actually the right way to do things, it
doesn't actually work at the moment but I might have to configure ibus
for linux to be able to receive the unicode.  I also need to confirm
the position of each letter on both the layers.